+++ CHANGELOG.md
@@ -3,9 +3,14 @@
 ## Unreleased
 
 - `nitrokey.trussed.admin_app.InitStatus`: add support for `EXT_FLASH_NEED_REFORMAT`
-- `nitrokey.nk3.updates`: return device status after an update
 - Use `poetry-core` v2 as build backend.
 - Bump minimum Python version to 3.10.
+- `nitrokey.trussed.Model`: Remove `firmware_repository` and `firmware_pattern` properties.
+- `nitrokey.nk3.updates`:
+  - Move to `nitrokey.trussed.updates` and prepare adding NKPK support.
+  - Return device status after an update.
+  - Add `model` argument to `get_firmware_update` and to the connection callbacks in `Updater`.
+  - Add `get_firmware_repository` method.
 
 [All Changes](https://github.com/Nitrokey/nitrokey-sdk-py/compare/v0.3.2...HEAD)